+++ libgpio/sources.cmake
@@ -1,11 +1,16 @@
 set(SOURCES
+	./include/libgpio/Device.hpp
+	./include/libgpio/DigitalOutput.hpp
 	./include/libgpio/LED.hpp
+	./include/libgpio/MotorDirection.hpp
+	./include/libgpio/MotorDriver.hpp
 	./include/libgpio/PullDirection.hpp
 	./include/libgpio/Switch.hpp
-	./include/libgpio/internal/Device.hpp
-	./src/libgpio/LED.cpp
+	./src/libgpio/Device.cpp
+	./src/libgpio/DigitalOutput.cpp
+	./src/libgpio/MotorDirection.cpp
+	./src/libgpio/MotorDriver.cpp
 	./src/libgpio/PullDirection.cpp
 	./src/libgpio/Switch.cpp
-	./src/libgpio/internal/Device.cpp
 )