+++ CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [1.0.0] - 2023-12-31
+## [1.2.0] - 2024-03-10
 
 ### Added
-- Examples showing how to control an LED using the multiple libaries (`pigpio`, `lgpio`, `bcm2835`, etc)
+- Examples showing how to control a motor when using the `pigpio` library functions
+- `MotorDriver` class to `libgpio` library which encapsulates the `pigpio` motor control functionality
+- `DigitalOutput` class to `libgpio` library which represents a generic digital output pin
+
+### Changed
+- Refactor `LED` class in `libgpio` library to use newly added`DigitalOutput` class
 
 
 ## [1.1.0] - 2024-02-15
@@ -19,3 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `libgpio` library that encapsulates the LED and Switch functionality using the `pigpio` library into C++ classes
 - Examples showing how to use the `libgpio` library to control an LED and read a switch
 - `CHANGELOG.md` file to keep track of changes in the project
+
+
+## [1.0.0] - 2023-12-31
+
+### Added
+- Examples showing how to control an LED using multiple libaries (`pigpio`, `lgpio`, `bcm2835`, etc)